+++ crates/aptos/src/move_tool/mod.rs
@@ -784,8 +784,7 @@ impl TryInto<PackagePublicationData> for &PublishPackage {
     type Error = CliError;
 
     fn try_into(self) -> Result<PackagePublicationData, Self::Error> {
-        let package =
-            build_package_options(&self.move_options, &self.included_artifacts_args).unwrap();
+        let package = build_package_options(&self.move_options, &self.included_artifacts_args)?;
 
         let package_publication_data =
             create_package_publication_data(package, PublishType::AccountDeploy, None)?;
@@ -815,8 +814,7 @@ impl AsyncTryInto<ChunkedPublishPayloads> for &PublishPackage {
     type Error = CliError;
 
     async fn async_try_into(self) -> Result<ChunkedPublishPayloads, Self::Error> {
-        let package =
-            build_package_options(&self.move_options, &self.included_artifacts_args).unwrap();
+        let package = build_package_options(&self.move_options, &self.included_artifacts_args)?;
 
         let chunked_publish_payloads =
             create_chunked_publish_payloads(package, PublishType::AccountDeploy, None)?;
@@ -1146,8 +1144,7 @@ impl CliCommand<TransactionSummary> for CreateObjectAndPublishPackage {
             let mock_object_address = AccountAddress::from_hex_literal("0xcafe").unwrap();
             self.move_options
                 .add_named_address(self.address_name.clone(), mock_object_address.to_string());
-            let package =
-                build_package_options(&self.move_options, &self.included_artifacts_args).unwrap();
+            let package = build_package_options(&self.move_options, &self.included_artifacts_args)?;
             let mock_payloads =
                 create_chunked_publish_payloads(package, PublishType::AccountDeploy, None)?
                     .payloads;
@@ -1162,8 +1159,7 @@ impl CliCommand<TransactionSummary> for CreateObjectAndPublishPackage {
         self.move_options
             .add_named_address(self.address_name, object_address.to_string());
 
-        let package =
-            build_package_options(&self.move_options, &self.included_artifacts_args).unwrap();
+        let package = build_package_options(&self.move_options, &self.included_artifacts_args)?;
         let message = format!(
             "Do you want to publish this package at object address {}",
             object_address
@@ -1246,7 +1242,7 @@ impl CliCommand<TransactionSummary> for UpgradeObjectPackage {
 
     async fn execute(self) -> CliTypedResult<TransactionSummary> {
         let built_package =
-            build_package_options(&self.move_options, &self.included_artifacts_args).unwrap();
+            build_package_options(&self.move_options, &self.included_artifacts_args)?;
         let url = self
             .txn_options
             .rest_options
@@ -1357,8 +1353,7 @@ impl CliCommand<TransactionSummary> for DeployObjectCode {
             let mock_object_address = AccountAddress::from_hex_literal("0xcafe").unwrap();
             self.move_options
                 .add_named_address(self.address_name.clone(), mock_object_address.to_string());
-            let package =
-                build_package_options(&self.move_options, &self.included_artifacts_args).unwrap();
+            let package = build_package_options(&self.move_options, &self.included_artifacts_args)?;
             let mock_payloads =
                 create_chunked_publish_payloads(package, PublishType::AccountDeploy, None)?
                     .payloads;
@@ -1373,8 +1368,7 @@ impl CliCommand<TransactionSummary> for DeployObjectCode {
         self.move_options
             .add_named_address(self.address_name, object_address.to_string());
 
-        let package =
-            build_package_options(&self.move_options, &self.included_artifacts_args).unwrap();
+        let package = build_package_options(&self.move_options, &self.included_artifacts_args)?;
         let message = format!(
             "Do you want to deploy this package at object address {}",
             object_address
@@ -1463,8 +1457,7 @@ impl CliCommand<TransactionSummary> for UpgradeCodeObject {
         self.move_options
             .add_named_address(self.address_name, self.object_address.to_string());
 
-        let package =
-            build_package_options(&self.move_options, &self.included_artifacts_args).unwrap();
+        let package = build_package_options(&self.move_options, &self.included_artifacts_args)?;
         let url = self
             .txn_options
             .rest_options