+++ crates/aptos/CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to the Aptos CLI will be captured in this file. This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and the format set out by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+- [`Fix`] Remove unwraps to make outputs go through regular error handling
 
 ## [4.2.5] - 2024/10/23
 - Bump to resolve issue with release version inconsistency.